fix: build issue and add component-based search functionality
