+++ app/docs/page.tsx
@@ -8,7 +8,7 @@ const MainDocsPage = () => {
 
   useEffect(() => {
     router.replace('/docs/overview/general');
-  }, []);
+  }, [router]);
 
   return null;
 };
