+++ components/HighlightedText/index.tsx
@@ -1,0 +1,30 @@
+// HighlightedText.tsx
+import React from 'react';
+
+const HighlightedText: React.FC<{ text: string; highlight: string }> = ({ text, highlight }) => {
+  if (!text || !highlight) {
+    return <span>{text}</span>;
+  }
+
+  // Escape special regex characters
+  const escapeRegExp = (string: string) => {
+    return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+  };
+
+  const parts = text.split(new RegExp(`(${escapeRegExp(highlight)})`, 'gi'));
+  return (
+    <span>
+      {parts.map((part, index) =>
+        part.toLowerCase() === highlight.toLowerCase() ? (
+          <mark key={index} className="bg-green-700 text-white">
+            {part}
+          </mark>
+        ) : (
+          part
+        )
+      )}
+    </span>
+  );
+};
+
+export default HighlightedText;
