+++ components/Navbar/index.tsx
@@ -5,7 +5,7 @@ import { Bars3Icon, XMarkIcon } from '@heroicons/react/24/outline'
 import Image from 'next/image';
 import Link from 'next/link';
 import { Dropdown } from '../Dropdown';
-import AskCookbook from '@cookbookdev/coreum/react';
+import AskCookbook from '@cookbookdev/docsbot/react';
 import "../index.css";
 
 // Import dynamic from Next.js
@@ -170,4 +170,3 @@ export const Navbar = () => {
     </>
   );
 }
-
